backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
pydantic
psycopg2-binary
django 
sentence-transformers
pythainlp
openai
httpx
tqdm
python-dotenv
loguru
transformers
torch
langchain
chromadb
scikit-learn
nltk
faker
